csv_agent/requirements.txt
# 
google-genai
langgraph
pandas
streamlit
python-dotenv
duckdb